fix: Watch items for updates
Note this is not a deep watch, just making sure we the correct reference to the
collection internally. We could just work with `$scope.items` but I think it
makes for a cleaner interface to toss these on the controller as well.

Also, remove the document click handler when $scope is $destroyed